third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCPeerConnection-connectionState-expected.txt

This is a testharness.js-based test.
FAIL Initial connectionState should be new assert_equals: expected (string) "new" but got (undefined) undefined
FAIL connection with one data channel should eventually have connected connection state assert_equals: Expect connection to have connectionstatechange event expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
